demo/requirements.txt
langchain==0.3.14
langchain-community==0.3.14
langchain-groq==0.2.4
langchain-core==0.3.28
faiss-cpu==1.9.0
requests==2.32.3
python-dotenv==1.0.1
